Refactor _ in package names
