+++ app/src/main/java/me/nikhilchaudhari/composeparticlesystem/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package me.nikhilchaudhari.compose_particle_system.ui.theme
+package me.nikhilchaudhari.composeparticlesystem.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
